split into logical parts: graph-rebuild / config-reading / plugin-implementation
