+++ plugin/src/main/java/com/github/mickroll/maven/dependency_duplicator_plugin/PluginLifecycleParticipant.java
@@ -1,6 +1,5 @@
 package com.github.mickroll.maven.dependency_duplicator_plugin;
 
-import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.LinkedHashMap;
@@ -15,29 +14,9 @@ import javax.inject.Inject;
 import org.apache.maven.AbstractMavenLifecycleParticipant;
 import org.apache.maven.MavenExecutionException;
 import org.apache.maven.execution.MavenSession;
-import org.apache.maven.execution.ProjectDependencyGraph;
-import org.apache.maven.lifecycle.internal.MojoDescriptorCreator;
 import org.apache.maven.model.Dependency;
-import org.apache.maven.model.Plugin;
-import org.apache.maven.plugin.InvalidPluginDescriptorException;
-import org.apache.maven.plugin.MojoNotFoundException;
-import org.apache.maven.plugin.PluginDescriptorParsingException;
-import org.apache.maven.plugin.PluginNotFoundException;
-import org.apache.maven.plugin.PluginResolutionException;
-import org.apache.maven.plugin.descriptor.MojoDescriptor;
-import org.apache.maven.plugin.prefix.NoPluginFoundForPrefixException;
-import org.apache.maven.plugin.version.PluginVersionResolutionException;
-import org.apache.maven.project.DuplicateProjectException;
 import org.apache.maven.project.MavenProject;
-import org.apache.maven.project.ProjectSorter;
-import org.codehaus.plexus.PlexusContainer;
 import org.codehaus.plexus.component.annotations.Component;
-import org.codehaus.plexus.component.configurator.BasicComponentConfigurator;
-import org.codehaus.plexus.component.configurator.ComponentConfigurationException;
-import org.codehaus.plexus.configuration.PlexusConfiguration;
-import org.codehaus.plexus.configuration.xml.XmlPlexusConfiguration;
-import org.codehaus.plexus.util.dag.CycleDetectedException;
-import org.codehaus.plexus.util.xml.Xpp3Dom;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -50,10 +29,10 @@ public class PluginLifecycleParticipant extends AbstractMavenLifecycleParticipan
     private static final Logger LOG = LoggerFactory.getLogger(PluginLifecycleParticipant.class);
 
     @Inject
-    private PlexusContainer container;
+    private PluginConfigResolver pluginConfigResolver;
 
     @Inject
-    private MojoDescriptorCreator mojoDescriptorCreator;
+    private DependencyGraphBuilder dependencyGraphBuilder;
 
     // needs maven 3.7.0, see https://github.com/apache/maven/pull/368
     // @Inject
@@ -75,106 +54,15 @@ public class PluginLifecycleParticipant extends AbstractMavenLifecycleParticipan
         addNewDependenciesToProjects(newProjectDependencies);
 
         LOG.info("rebuilding project dependency graph");
-        rebuildDependencyGraph(session);
+        dependencyGraphBuilder.rebuildDependencyGraph(session);
 
         LOG.info("finished after {}ms.", System.currentTimeMillis() - startTime);
-    }
-
-    private Optional<PluginMojo> findPluginConfig(final MavenSession session, final MavenProject project) {
-        MavenProject projectCursor = project;
-        Plugin plugin = null;
-        while (projectCursor != null) {
-            plugin = projectCursor.getPlugin(PluginMojo.PLUGIN_KEY);
-            if (plugin != null) {
-                // break to keep correct projectCursor
-                break;
-            }
-            projectCursor = projectCursor.getParent();
-        }
-
-        if (plugin == null) {
-            LOG.debug("plugin definition not found in project or parent projects");
-            return Optional.empty();
-        }
-        if (plugin.getConfiguration() == null) {
-            LOG.debug("plugin configuration not found in project or parent projects");
-            return Optional.empty();
-        }
-
-        try {
-            final MojoDescriptor descriptor = mojoDescriptorCreator.getMojoDescriptor(PluginMojo.PLUGIN_KEY + ":" + PluginMojo.GOAL, session, projectCursor);
-            LOG.debug("descriptor created " + descriptor);
-
-            descriptor.setConfiguration(getConfig(plugin.getConfiguration()));
-            final PlexusConfiguration config = getConfig(plugin.getConfiguration());
-
-            final PluginMojo mojo = new PluginMojo();
-            new BasicComponentConfigurator().configureComponent(mojo, config, container.getContainerRealm());
-
-            return Optional.of(mojo);
-        } catch (PluginNotFoundException | PluginResolutionException | PluginDescriptorParsingException | MojoNotFoundException
-                | NoPluginFoundForPrefixException | InvalidPluginDescriptorException | PluginVersionResolutionException
-                | ComponentConfigurationException e) {
-            LOG.error("unable to read plugin configuration ", e);
-            return Optional.empty();
-        }
-    }
-
-    private PlexusConfiguration getConfig(final Object config) {
-        if (config instanceof PlexusConfiguration) {
-            return (PlexusConfiguration) config;
-        } else if (config instanceof Xpp3Dom) {
-            return new XmlPlexusConfiguration((Xpp3Dom) config);
-        }
-        throw new IllegalArgumentException("unable to convert config to plexus config: " + config);
-    }
-
-    private void rebuildDependencyGraph(final MavenSession session) {
-        final ProjectDependencyGraph graph = session.getProjectDependencyGraph();
-        final ProjectSorter newSorter;
-        try {
-            newSorter = new ProjectSorter(graph.getAllProjects());
-        } catch (CycleDetectedException | DuplicateProjectException e) {
-            LOG.error("unable to rebuild project dependency graph", e);
-            return;
-        }
-
-        if ("org.apache.maven.graph.DefaultProjectDependencyGraph".equals(graph.getClass().getName())) {
-            replaceProjectSorter(graph, "sorter", newSorter);
-        } else if ("org.apache.maven.graph.FilteredProjectDependencyGraph".equals(graph.getClass().getName())) {
-            try {
-                final Field field = graph.getClass().getDeclaredField("projectDependencyGraph");
-                field.setAccessible(true);
-                final Object nestedGraph = field.get(graph);
-                if ("org.apache.maven.graph.DefaultProjectDependencyGraph".equals(nestedGraph.getClass().getName())) {
-                    replaceProjectSorter(nestedGraph, "sorter", newSorter);
-                } else {
-                    LOG.warn(
-                            "unable to rebuild project dependency graph, unexpected nested graph implementation found in {}: {}",
-                            graph.getClass().getName(), nestedGraph.getClass().getName());
-                }
-            } catch (IllegalArgumentException | IllegalAccessException | NoSuchFieldException | SecurityException e) {
-                LOG.error("unable to set rebuilt project dependency graph", e);
-            }
-        } else {
-            LOG.warn("unable to rebuild project dependency graph, unexpected graph implementation found: {}", graph.getClass().getName());
-        }
-    }
-
-    private void replaceProjectSorter(final Object obj, final String fieldName, final ProjectSorter newSorter) {
-        try {
-            final Field field = obj.getClass().getDeclaredField(fieldName);
-            field.setAccessible(true);
-            field.set(obj, newSorter);
-        } catch (IllegalArgumentException | IllegalAccessException | NoSuchFieldException | SecurityException e) {
-            LOG.error("unable to set rebuilt project dependency graph", e);
-        }
     }
 
     private Map<MavenProject, DependencySet> createDuplicateDependenciesForProjects(final MavenSession session) {
         final Map<MavenProject, DependencySet> newProjectDependencies = new LinkedHashMap<>();
         for (final MavenProject project : session.getAllProjects()) {
-            final Optional<PluginMojo> pluginConfig = findPluginConfig(session, project);
+            final Optional<PluginMojo> pluginConfig = pluginConfigResolver.findPluginConfig(session, project);
             if (!pluginConfig.isPresent()) {
                 LOG.info("[{}] did not find any plugin config", project.getName());
                 continue;
